Rename stub audio manager file so that it doesn't clash with the audio manager test file.
