+++ test/test_engine.js
@@ -4,7 +4,7 @@ import app from "../server.mjs";
 import GameEngine from "../public/js/engine.js"; 
 import {Player} from "../public/js/player.js";
 import {Platform} from "../public/js/platform.js";
-import {TestAudioManager} from "./test_audio_manager.js";
+import {TestAudioManager} from "./stub_audio_manager.js";
 
 let server;
 var TestContext = require("./context.js");
